feat: llm inference optimization
